+++ insert_test.go
@@ -20,8 +20,7 @@ func TestInsertStmt(t *testing.T) {
 	}).Comment("INSERT TEST")
 	err := builder.Build(dialect.MySQL, buf)
 	require.NoError(t, err)
-	require.Equal(t, "INSERT INTO `table` (`a`,`b`) VALUES (?,?), (?,?)", buf.String())
-	//  	require.Equal(t, "/* INSERT TEST */\nINSERT INTO `table` (`a`,`b`) VALUES (?,?), (?,?)", buf.String())
+	require.Equal(t, "/* INSERT TEST */\nINSERT INTO `table` (`a`,`b`) VALUES (?,?), (?,?)", buf.String())
 	require.Equal(t, []interface{}{1, "one", 2, "two"}, buf.Value())
 }
